오류수정
오류수정 개꿀띠

+++ Assets/Scripts/World/GameManager/GameManagerScript.cs
@@ -67,7 +67,7 @@ public class GameManagerScript : MonoBehaviour
 
         if(canConvert){
         total_money = result;
-        moneyindicator.text = numberManager.ToCurrencyString(total_money);
+        moneyindicator.text = numberManager.ToCurrencyString((int)total_money);
 
         money_per_seconds = normal_money;
         fever_money_per_seconds *= normal_money; 
@@ -85,7 +85,7 @@ public class GameManagerScript : MonoBehaviour
         //돈늘어나는거 1초마다 
         if(Time.time> nextPayMoney){
         total_money += money_per_seconds;
-        moneyindicator.text = numberManager.ToCurrencyString(total_money);
+        moneyindicator.text = numberManager.ToCurrencyString((int)total_money);
         nextPayMoney = Time.time+ interval;
         }
         if(Input.GetKeyDown(KeyCode.Escape)){
